選択文字列が空のデータを置換するように修正
